Refactor Countries and create Country component
