+++ src/styling/index.css
@@ -1,3 +1,6 @@
+*{
+  box-sizing: border-box;
+}
 body {
   margin: 0;
   font-family:
@@ -24,4 +27,10 @@ code {
     Consolas,
     'Courier New',
     monospace;
+}
+
+ul  {
+  list-style: none;
+  margin: 0;
+  padding: 0 !important;
 }
